Now that pr 688 has been merged, these old tests work.
Stemming from a lack of documentation, I didn't realize the patch 400 error was happening because policy tags were being applied (but only an empty list of them!) to non leaf fields, namely the inner struct (which itself has an inner struct with leaf fields). In pr 542, this bug was introduced -- but why did it initially pass CI? That's a mystery I have no answer for: Guess is a cloud change that exposed this or a BQ spec change.