fix: cast dataproc batch runtime props
There is a bug currently that requires `runtime_config` `properties` values to
be strings. As a result, integer values need quotes around them in the yml
files, which is awkward.

With this PR, values are cast as strings, so that the properties can be set in
a more expected way.

[ADAP-1007]